Solve linters error
